change  file names, deleted  import react, deleted heading elements
