+++ src/App.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import {
     Header,
     HeroSection,
