--- src/components/DownloadSection/DownloadSectionContent/DownloadSectionText/DownloadSectionText.tsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import H2 from "@UI/Text/H2";
-import H4 from "@UI/Text/H4";
-
-import './DownloadSectionText.scss';
-
-
-const DownloadSectionText = () => {
-    return <div className="download-section__text">
-        <H2>Download the app now.</H2>
-        <H4>Available on your favorite store. Start your premium experience now</H4>
-    </div>
-}
-
-export default DownloadSectionText;
