+++ src/components/DownloadSection/DownloadSectionContent/DownloadSectionText/index.scss
