+++ src/components/DownloadSection/DownloadSectionContent/DownloadSectionText/index.tsx
@@ -1,0 +1,11 @@
+import './index.scss';
+
+
+const DownloadSectionText = () => {
+    return <div className="download-section__text">
+        <h2>Download the app now.</h2>
+        <h4>Available on your favorite store. Start your premium experience now</h4>
+    </div>
+}
+
+export default DownloadSectionText;
