+++ src/components/DownloadSection/DownloadSectionContent/index.scss
