+++ src/components/Footer/Social/Icons/Facebook.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import './Icons.scss'
 
 const Facebook = () => {
