+++ src/components/Footer/Social/Icons/Linkedin.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 import './Icons.scss'
 
 const Linkedin = () => {
