+++ src/components/Footer/Social/index.scss
