+++ src/components/Footer/Social/index.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
-
-import './Social.scss'
 import Twitter from "./Icons/Twitter";
 import Facebook from "./Icons/Facebook";
 import Linkedin from "./Icons/Linkedin";
+
+import './index.scss'
+
 
 const Social = () => {
     return <div className="social">
