+++ src/components/Footer/index.scss
