+++ src/components/HeroSection/HeroSectionContent/HeroSectionImage/index.scss
