--- src/components/HeroSection/HeroSectionContent/HeroSectionText/HeroSectionText.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import H1 from '@UI/Text/H1';
-import H3 from '@UI/Text/H3';
-import H4 from '@UI/Text/H4';
-
-import './HeroSectionText.scss';
-
-const HeroSectionText = () => {
-    return <div className="hero-section__text">
-        <H3>Food app</H3>
-        <H1>Why stay hungry when
-            you can order form Bella Onojie</H1>
-        <H4>Download the bella onoje’s food app now on</H4>
-    </div>
-}
-
-export default HeroSectionText;
