+++ src/components/HeroSection/HeroSectionContent/HeroSectionText/index.scss
