+++ src/components/HeroSection/HeroSectionContent/HeroSectionText/index.tsx
@@ -1,0 +1,12 @@
+import './index.scss';
+
+const HeroSectionText = () => {
+    return <div className="hero-section__text">
+        <h3>Food app</h3>
+        <h1>Why stay hungry when
+            you can order form Bella Onojie</h1>
+        <h4>Download the bella onoje’s food app now on</h4>
+    </div>
+}
+
+export default HeroSectionText;
