+++ src/components/HeroSection/HeroSectionContent/index.scss
