+++ src/components/HeroSection/index.scss
