+++ src/components/HowWorks/Step/StepText/index.scss
