+++ src/components/HowWorks/Step/StepText/index.tsx
@@ -1,10 +1,5 @@
-import React from 'react';
 
-import H2 from "@UI/Text/H2";
-import H3 from "@UI/Text/H3";
-import H4 from "@UI/Text/H4";
-
-import './StepText.scss';
+import './index.scss';
 
 interface stepTextProps {
     h2: string
@@ -14,9 +9,9 @@ interface stepTextProps {
 
 const StepText = ({h2, h3, h4}: stepTextProps) => {
     return <div className="howworks-section__text">
-        <H3>{h3}</H3>
-        <H2>{h2}</H2>
-        <H4>{h4}</H4>
+        <h3>{h3}</h3>
+        <h2>{h2}</h2>
+        <h4>{h4}</h4>
     </div>
 }
