+++ src/components/HowWorks/Step/index.scss
