--- src/components/HowWorks/Title/Title.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import './Title.scss'
-import H1 from "@UI/Text/H1";
-
-const Title = () => {
-    return <div className="howworks-section__title">
-        <H1>How the app works</H1>
-    </div>
-}
-
-export default Title;
