+++ src/components/HowWorks/Title/index.scss
