+++ src/components/HowWorks/Title/index.tsx
@@ -1,0 +1,9 @@
+import './index.scss'
+
+const Title = () => {
+    return <div className="howworks-section__title">
+        <h1>How the app works</h1>
+    </div>
+}
+
+export default Title;
