+++ src/components/HowWorks/index.scss
