+++ src/components/UI/Image/index.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-
 interface imageProps {
     src: string
     alt: string
