--- src/components/UI/Text/H1.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import './Text.scss'
-
-import {textProps} from "./textProps";
-
-const H1 = ({children}: textProps) => {
-    return <h1>{children}</h1>
-}
-
-export default H1;
