--- src/components/UI/Text/H4.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import './Text.scss'
-
-import {textProps} from './textProps';
-
-const H4 = ({children}: textProps) => {
-    return <h4>{children}</h4>
-}
-
-export default H4;
