--- src/components/UI/Text/textProps.ts
@@ -1,4 +1,0 @@
-export interface textProps {
-    children: string
-    id? : string
-}
