+++ src/index.scss
@@ -1,0 +1,51 @@
+@import 'Styles/mediaSizes';
+
+*{
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+
+h1{
+  font-size: 58px;
+  font-weigth: 700;
+  margin: 11px 0;
+}
+
+h2{
+  font-size: 40px;
+  font-weigth: 700;
+}
+
+h3{
+  font-size: 24px;
+  font-weigth: 700;
+  margin: 6px 0;
+}
+
+h4{
+  font-size: 24px;
+  font-weight: 500;
+  margin: 16px 0;
+}
+
+p {
+  font-size: 12px;
+  font-weight: 300;
+}
+
+@media (max-width: $breakpoint-mobile) {
+  h1 {
+    font-size: 36px;
+  }
+}
+
+
