+++ src/index.tsx
@@ -1,6 +1,5 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
+import './index.scss';
 import App from './App';
 
 ReactDOM.render(<App />,  document.getElementById('root'));
